backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
pillow==10.4.0
qrcode[pil]==7.4.2
gunicorn==23.0.0
psycopg2-binary==2.9.10
